+++ src/chinodeco/debug/__init__.py
@@ -1,2 +1,13 @@
 # !/usr/bin/env Python3
 # -*- coding:utf-8 -*-
+
+from .debugger import (
+    DEBUG,
+    debug
+)
+
+from .errors import (
+    UnknownCommandError,
+    UnknownParameterError,
+    ArgumentCountError
+)