+++ src/chinodeco/debug/errors.py
@@ -1,2 +1,11 @@
 # !/usr/bin/env Python3
 # -*- coding:utf-8 -*-
+
+class UnknownCommandError(Exception):
+    pass
+
+class UnknownParameterError(Exception):
+    pass
+
+class ArgumentCountError(Exception):
+    pass