+++ src/chinodeco/decodsl/__init__.py
@@ -4,3 +4,7 @@
 from .control import (
     when
 )
+
+from .registry import (
+    CommandDispatcher
+)